BUG: do not conjoin init in SLUGSIN
because the structured SLUGS format is infix, so an "&" in the middle
works. In contrast, SLUGSIN is a prefix format, so an "&" would have to
be added as prefix. This is not necessary, since multiple lines in init
are conjoined by SLUGS (this convention is common to other synthesizers
too).