instantiate parser with different AST, instead of subclassing
because overriding a production triggers table generation, so an extra layer of complexity.

This draws the border in the trade-off between reuse of a grammar (in the form of a parser) and the associated class hierarchy that the parser knows (so not refining the AST node classes), vs adding case selection statements to each flattener.

As long as the selection is among calls to clearly named and focused flattener functions, readability is preserved, and indentation decreased.

Note that the parser instance is hidden, whereas the nodes are not. This ensures that the global var `_parser` is not exposed, but the `Nodes` are, so that the same ingredients become available also outside `bitvector`, to instantiate similar parsers there.