BUG: recognize both prime and `X` as priming of vars
`X` is a Promela operator, so a `Unary` instance is instantiated by `promela.yacc`.
Priming is an open Promela operator, and a `Next` instance is instantiated (a subclass of `Unary`).
However, only `Next` was collected in `collect_primed_vars` (a consequence of too complicated an AST class hierarchy).

Fix: duck test for the desired operator.