BUG: freeze env when atomic request pending
Otherwise the environment can change a global free variable to block the system and so force it to lose atomicity. But when the system has requested atomic execution (source node context = atomic), it loses atomicity only if it blocks alone, assuming the rest of the world doesn't change. If the environment is allowed to interrupt and change something to block the system in the middle of an atomic block, then that violates the request of atomicity.

In other words, there is a priority of evaluation here. If the environment remains frozen implies that the system will not block in the next step of its atomic block, then the environment is obliged to grant the request for atomic execution. Otherwise, (if the system blocks despite env remaining frozen), then the environment interrupts the atomic block, and the system loses atomicity.