add proctypes to symbol table, instead of a list
split the instantiation of processes. First, all process instances are added to the symbol table. The translation happens in a second pass, so that remote reference statements will be able to find instances by pid (if a proctype has more than one instances), or find the pid of the single instance (if a proctype has a single instance).

For this reason, split `add_processes`, with the flattening moved to the function `processes_to_logic`. Extract the process instantiation (pid creation) from `process_to_logic` to `add_processes`.