mv `bitvector` and `bdd` modules to package `omega`
together with their tests. The reason is that the bitblaster is used extensively by other code, for example `omega.symbolic.symbolic`. The `openpromela` package depends on `omega`. If modules of the latter depend on modules of `openpromela`, then this results in a circular dependency.

These modules are stable and self-contained. By placing them in `omega`, they are "pushed" lower in the python stack for reasoning about omega-regular sets.

destination: https://github.com/johnyf/omega