BUG: redefine `grantee_lockreq` in AMBA
The variable `grantee_lockreq` abstracts the bit `lockreq[grant']` of the environment array `lockreq` in the original specification (HLOCK). This substitution of `grantee_lockreq` and `master_lockreq` in place of the two elements `lockreq[grant]` and `lockreq[master]` is a form of variable hiding (hiding of the `lockreq[i]` bits controlled by the masters from the arbiter).

It is impossible to construct in hardware the input signal `grantee_lockreq` from the signals `lockreq[grant']`, because the value `grant'` is selected by the arbiter after the value of the `lockreq[i]` signals. This sequencing is invariant under time refinement. So it holds in games with both Moore and Mealy time refinement.

The amendment is to shift the setting of `grantee_lockreq` by one time step, so that the signal can be constructed as `grantee_lockreq' = (--X lockreq)[grant]`. By introducing priming for `lockmemo`, the lock request is still propagated in the next time step to `lock memo`, and to `lock`, if a change of master happens in the next time step.