BUG: prime `master_lockreq` in AMBA
because the environment cannot otherwise set `master_lockreq` to the value of `lockreq[master']`, as it does not know the value of `master'` yet. Note that this is an artifact of importing the priming from a Moore to a Mealy game. It is not a sequencing in time that is invariant under time refinement, as in the case of the `grantee_lockreq` variable.