attach array-theory info to bits
so that an informed selection can be made later for the initial variable order. This is essential for when variable reordering is not dynamic. As of v0.0.5, `dd.bdd` offers static reordering, so it is vulnerable to a bad variable order. CUDD is not, so `dd.cudd` can be used in the future, when computing guards.

A bad variable order can have an exponentially bad effect in absence of variable reordering. This was observed during computation of guard conditions that involved a bit array, and an integer variable as index of array elements. In the variable order, the index must appear *above* the array elements. If all arrays are of Boolean type, then this can be ensured by placing all array elements at the bottom of the variable order.

In the future, the variable ordering will be selected by utilizing more information available in the syntactic structure of the specification.