BUG: negate persistence properties
Introduced in 44fa74982536f7de6804658021148f12c8564489.
This error has its mirror in `openpromela.slugs._to_slugs`.
The mirror passed the <>[] values as environment recurrence to `slugs`.
The result was correct, because the <>[] values were equal to the environment recurrence goals (due to the forgotten negation).
This commit turns the original omission of negation into a bug.